do not delete unexistent files

+++ src/Imap2AF/Mailboxer.php
@@ -197,7 +197,9 @@ class Mailboxer extends Mailbox
     {
         parent::__destruct();
         foreach ($this->filesToClean as $fileToClean) {
-            unlink($fileToClean);
+            if (file_exists($fileToClean)) {
+                unlink($fileToClean);
+            }
         }
     }
 }
